pyproject.toml
[project]
name = "saas-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.17.2",
    "fastapi>=0.123.9",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]

